+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Fixed multiple nested folders:
+  https://github.com/VadimNastoyashchy/real-test-js/issues/59.
+
 ## 1.6.0 - 2024-01-12
 ### Added
 - New features: